A Hideously Functional⁽™⁾ ecs-system
Entities are vanilla lua tables whose components are its fields /
table-keys.

A System is code which acts upon entities based on whether they match
its filter; which are compared to the entity components in various ways.

Currently not much happens, except filtering...

Dependencies are installed with luarocks (and will be added in the next
commit)

Testing is done with luaunit
Some generic utils live in their own file.